samples/agent/adk/generic_chat/pyproject.toml
[project]
name = "a2ui-generic-chat"
version = "0.1.0"
description = "Generic A2UI Chat Agent"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "a2a-sdk>=0.3.0",
    "click>=8.1.8",
    "google-adk>=1.8.0",
    "google-genai>=1.27.0",
    "python-dotenv>=1.1.0",
    "litellm",
    "jsonschema>=4.0.0",
    "uvicorn>=0.20.0",
    "fastapi>=0.95.0",
    "a2ui",
]

[tool.hatch.build.targets.wheel]
packages = ["."]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.metadata]
allow-direct-references = true
